refactor: extract general-Go naming helpers to extraction/go/naming
split_receiver and normalize_go_entry_point are general Go naming
rules (method spelling, runtime -fm/.funcN decorations), not
Terraform-specific and not tied to the model-generation feature.
Move them to an always-compiled extraction/go/naming module and have
the Terraform plan consumer (go_symbol) delegate to them. No behavior
change.
